requirements.txt
allure-pytest
jsonref
jsonschema
requests
pytest
pytest-xdist
jsonref
jsonschema
pytest-order
